requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==4.2.16
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
django-model-utils==5.0.0
django-nine==0.2.7
django-notifications-hq==1.8.3
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
jsonfield==3.1.0
msgpack==1.0.8
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
python-decouple==3.8
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytest==8.3.2
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
ruff==0.6.4
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
swapper==1.4.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
zope.interface==7.0.3
